Ai-model-medic/requirements.txt
Flask
scipy
scikit-learn
joblib
google-generativeai
python-dotenv
